JSON import: added event types and events, better exception handling

+++ wsgi/openshift/confla/views.py
@@ -14,7 +14,7 @@ from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext as _
 from django.utils import timezone
 
-from confla.models import ConflaUser, Conference, Room, Timeslot, EmailAdress
+from confla.models import * 
 from confla.forms import *
 
 class AboutView(generic.TemplateView):
@@ -316,7 +316,7 @@ class ImportView(generic.TemplateView):
                         "name" : "",
                         "desc" : "",
                         "color" : ""
-                }],
+                    }],
                 "confs" : [{
                         "id" : 100,
                         "name" : "JSON_conf",
@@ -327,14 +327,21 @@ class ImportView(generic.TemplateView):
                         "end_d" : "28/03/2016",
                         "time_d" : 10,
                         "active" : true
-                }],
+                    }],
                 "slots" : [{
                         "id" : 1,
                         "conf_id" : 100,
                         "room_id" : 100,
                         "start_t" : "10:10",
                         "end_t" : "10:50"
-                }],
+                    },
+                    {
+                        "id" : 2,
+                        "conf_id" : 100,
+                        "room_id" : 100,
+                        "start_t" : "12:10",
+                        "end_t" : "12:50"
+                    }],
                 "users" : [{
                         "username" : "testor",
                         "f_name" : "First name",
@@ -350,7 +357,23 @@ class ImportView(generic.TemplateView):
                         "g+" : "",
                         "linkedin" : "",
                         "bio" : ""
-                }]
+                    }],
+                "e_types" : [{
+                         "id" : 1,
+                         "name" : "talk"
+                    }],
+                "events" : [{
+                         "id" : 1,
+                         "type_id" : 1,
+                         "conf_id" : 100,
+                         "topic" : "topic",
+                         "desc" : "description",
+                         "lang" : "Czech",
+                         "slides" : "",
+                         "video" : "",
+                         "g_doc" : "",
+                         "speakers" : ["testor"]
+                    }]
             }"""    
 
     def json_to_db(json_string):
@@ -367,9 +390,13 @@ class ImportView(generic.TemplateView):
             newroom.color = room['color']
             try:
                 newroom.full_clean()
-            except ValidationError:
-                Room.objects.get(id=newroom.id).delete()
-                newroom.full_clean()
+            except ValidationError as e:
+                if ('id' in e.error_dict and
+                        e.error_dict['id'][0] == 'Room with this ID already exists.'):
+                    Room.objects.get(id=newroom.id).delete()
+                    newroom.full_clean()
+                else:
+                    raise e
             newroom.save()
 
         # Generate conferences
@@ -392,9 +419,14 @@ class ImportView(generic.TemplateView):
             # Conference has to be in db before we can add rooms
             try:
                 newconf.full_clean()
-            except ValidationError:
-                Conference.objects.get(id=newconf.id).delete()
-                newconf.full_clean()
+            except ValidationError as e:
+                if ('id' in e.error_dict and
+                        e.error_dict['id'][0] == 'Conference with this ID already exists.'):
+                    Conference.objects.get(id=newconf.id).delete()
+                    newconf.full_clean()
+                else:
+                    raise e
+
             newconf.save()
 
             for room in conf['rooms']:
@@ -417,9 +449,13 @@ class ImportView(generic.TemplateView):
                                                         second=0, microsecond=0)
             try:
                 newslot.full_clean()
-            except ValidationError:
-                Timeslot.objects.get(id=newslot.id).delete()
-                newslot.full_clean()
+            except ValidationError as e:
+                if ('id' in e.error_dict and
+                        e.error_dict['id'][0] == 'Timeslot with this ID already exists.'):
+                    Timeslot.objects.get(id=newslot.id).delete()
+                    newslot.full_clean()
+                else:
+                    raise e
             newslot.save()
 
         # Generate users
@@ -443,7 +479,55 @@ class ImportView(generic.TemplateView):
             newuser.bio = user['bio']
             try:
                 newuser.full_clean()
-            except ValidationError:
-                ConflaUser.objects.get(username=newuser.username).delete()
-                newuser.full_clean()
+            except ValidationError as e:
+                if ('username' in e.error_dict and
+                        e.error_dict['username'][0] == 'Confla user with this Username already exists.'):
+                    ConflaUser.objects.get(username=newuser.username).delete()
+                    newuser.full_clean()
+                else:
+                    raise e
             newuser.save()
+
+        # Generate event types
+        event_type_list = json_obj['e_types']
+        for etype in event_type_list:
+            newtype = EventType()
+            newtype.id = etype['id']
+            newtype.name = etype['name']
+            try:
+                newtype.full_clean()
+            except ValidationError as e:
+                if ('id' in e.error_dict and
+                        e.error_dict['id'][0] == 'Event type with this ID already exists.'):
+                    EventType.objects.get(id=newtype.id).delete()
+                    newtype.full_clean()
+                else:
+                    raise e
+            newtype.save()
+
+        # Generate events
+        event_list = json_obj['events']
+        for event in event_list:
+            newevent = Event()
+            newevent.id = event['id']
+            newevent.conf_id = Conference.objects.get(id=event['conf_id'])
+            newevent.e_type_id = EventType.objects.get(id=event['type_id'])
+            newevent.topic = event['topic']
+            newevent.description = event['desc']
+            newevent.lang = event['lang']
+            newevent.slides = event['slides']
+            newevent.video = event['video']
+            newevent.google_doc_url = event['g_doc']
+            try:
+                newevent.full_clean()
+            except ValidationError as e:
+                if ('id' in e.error_dict and
+                        e.error_dict['id'][0] == 'Event with this ID already exists.'):
+                    Event.objects.get(id=newevent.id).delete()
+                    newevent.full_clean()
+                else:
+                    raise e
+            newevent.save()
+            for speaker in event['speakers']:
+                newevent.speaker.add(ConflaUser.objects.get(username=speaker))
+            newevent.save()
